Rename related targets Detail
